using int rather than uin64_t for emitted signal - used to work but seem to be qt version dependent (#260)

+++ source/daqlite/WorkerThread.cpp
@@ -30,7 +30,7 @@ void WorkerThread::run() {
       Consumer->mHistogramTofPlot = Consumer->mHistogramTof;
       mutex.unlock();
 
-      uint64_t ElapsedCountMS = elapsed.count()/1000000;
+      int ElapsedCountMS = elapsed.count()/1000000;
       emit resultReady(ElapsedCountMS);
 
       std::fill(Consumer->mHistogram.begin(), Consumer->mHistogram.end(), 0);
